(refactor): extract duplicate remover content component
